Add files via upload
Ts to build a machine learning model capable of accurately classifying tweets
as either related to real disasters or not. While some tweets may contain explicit keywords
like "fire" or "flood," others may use such terms metaphorically or in a non-disaster context.
Thus, the model needs to discern the true intent behind the language used in tweets, which
can be nuanced and context-dependent.



Part 1: Data Exploration and Preparation


Part 2: Feature Engineering and Model Selection


Part 3: Model Evaluation and Validation

Part 4: Deployment with Web Interface


we can enhance the ability to detect and respond to emergencies more
efficiently. This project represents a valuable opportunity to leverage NLP techniques for the
greater good, contributing to the advancement of disaster management and public safety
efforts.